fix(hub): lazy-load team access and bound submissions
